+++ src/core/types.ts
@@ -199,6 +199,7 @@ export interface ContextInterface {
   renderPlaceholder?: ((context: Context) => BoxInterface) | undefined;
   readonly smartFence: boolean;
   readonly letterShapeStyle: 'tex' | 'french' | 'iso' | 'upright';
+  readonly minFontScale: number;
   readonly placeholderSymbol: string;
   readonly colorMap: (name: string) => string | undefined;
   readonly backgroundColorMap: (name: string) => string | undefined;